+++ frontend/packages/erd-core/src/components/ERDRenderer/ERDContent/index.ts
@@ -1,3 +1,3 @@
 export * from './ERDContent'
 export * from './TableNode'
-export * from './Toolbar'
+export * from './useAutoLayout'
