+++ frontend/packages/erd-core/src/components/ERDRenderer/Toolbar/TidyUpButton/index.ts
