+++ frontend/packages/erd-core/src/components/ERDRenderer/Toolbar/ZoomControls/ZoomControls.module.css
